ObjKitT in annotation inference
Summary:
Similar to Flow_js, operations over EvalTs cause the EvalT to be evaluated before
the operation can be applied. In annotation inference, we are not be able to evaluate
any EvalT; specifically, we are only able to handle cases that don't involve
subtyping ($Call, etc.).

This diff handles cases that rely on Slice_utils and instantiates it appropriately.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D29422816

fbshipit-source-id: 10be665f7ac6edf793d4386f3507584dffd80d69