ObjRestT (partial) in annotation inference
Summary:
This adds support for processing ObjRestT constraints in annotation inference
for all cases except InstanceT inputs. The logic for supported cases is shared
with flow_js.ml. The case of InstanceT is significantly more complicated, since it
relies on unsealed objects and set-prop logic. Instead, we error with an "unsupported"
error.

Reviewed By: mvitousek

Differential Revision: D31696864

fbshipit-source-id: 635af96d533d785d1e66504b5c020c026aa79bb6