$Keys in annotation inference
Summary:
It is not common to need evaluate $Keys during merge, but when it does
occur when a subsequent operation forces us. For example,
```
declare var k: $Keys<typeof obj>;
export type T = typeof k.length;
```
This diff adds support for `$Keys` in annotation inference. There is one small difference
in implementation compared to the one in Flow_js: Instead of flowing the
result of the `ToStringT` operation as lower bounds to a `keys` tvar that collects
the result, we create a union of the all the key types.

Reviewed By: samwgoldman

Differential Revision: D31374428

fbshipit-source-id: f3c739c6e34db380e2069a2d1e858006519ec9e6