error reporting in annotation inference
Summary:
In types-first 2.0, converting type-sigs to types (Type.t) happens during the check
phase. (This used to happen during the merge phase in types-first 1.0.) The context
in which these conversions happen is the one of the originating file of the definitions.
Errors that are raised during this phase are not reported, since they are not recorded
in the context of the file being checked. The intuition behind this decision is that these
errors are already reported as part of the checking phase of these origin files.

Note that in the existing version (non-annotation inference), even though errors are
not reported, they are still important to be raised, as mechanisms such as speculation
rely on them to function.

Annotation inference changes these facts in two ways:
* There are certain kinds of errors that need to be reported during this type-sig to
Type.t conversion. These are errors equivalent to "signature-verification failures",
ie type definitions that are not supported in an export position. These cases are
rather rare, but need to be reported nonetheless.
* Annotation inference does not rely on the regular error reporting that used to
happen during this phase: (i) There is no speculation, and (ii) no support for
operations on call-like utility types. At this point, we still raise these errors, mostly
due to code reuse with flow_js.ml. In the future these errors can be cleaned up.

So, this diff enables the above by:
* Providing Annotation_inference with access to the destination context for error
reporting. This happens through a reference `dst_cx_ref` to that context as soon
as it created in check_service.ml.
* Creating a new `EAnnotationInference` error to represent unsupported
exported types. These are mostly *operations* on call-like types, maybe type, and
idx-wrapper (see instances of `error_unsupported`).
* Creating specialized reasons for annot_ops (`display_reason_of_op`) for
better error reporting.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D31678448

fbshipit-source-id: 47609b39db13aa9936732c94209c9cd56b6eb9eb