[codemods] fix interplay of max-type-size and default-any flags
Summary:
Before, `validate_ty` used to return `Err (errs, ty)`, where `ty` was the type
under check, in case of a violation. In the non-`default-any` case, `ty` was never
used when a violation occurred. But with `default-any` set, this type would
be used directly, instead of `Ty.explicit_any` which would be more appropriate.
This diff fixes this by removing the `ty` part of the `Error` output. The `any` type
constructor is moved to the part where the `default_any` case is checked.

Reviewed By: evanyeung

Differential Revision: D31812587

fbshipit-source-id: 566606e79dcd6a3538f4e8c335c5383d9ae5f7ee