remove global find refs and renaming
Summary:
global find refs (`textDocument/references` and `flow find-refs --global`) has been in "beta" for years. sometimes it works, sometimes it takes a super long time, sometimes it crashes the server (e.g. using the workers when the workers are already busy). similar for `textDocument/rename`, which is powered by the same code.

we don't have the bandwidth to improve this -- or even maintain it, really -- so we're removing it for the time being. we do recognize that this is an important feature, and will bring it back as soon as we can.

since we're no longer advertising the `referencesProvider` or `renameProvider` LSP capabilities, this shouldn't be a "breaking" change, per se, when using the LSP. `flow find-refs --global` is now an error.

Changelog: [feat?] Support for finding references and renaming has been removed. While these features often worked well, they also caused extremely bad performance and crashes. We hope to reimplement these features as soon as we can.

Reviewed By: vrama628

Differential Revision: D31829851

fbshipit-source-id: ce5e8bd2289b9f735a9d206d9cfac189bd4a07d7