Do not re-visit assignment LHS
Summary:
We are redundantly re-visiting the LHS of an assignment in conditionals here and asking for the _read_ value of the LHS from the EnvBuilder. The LHS is already visited in the call to `expression` in this match case, so visiting this LHS again is unnecessary.

Furthermore, visiting this expression in this manner causes the EnvBuilder to crash. The LHS of an assignment is not a read of a variable, but traversing it as if it was just an identifier as an expression will query the EnvBuilder for the read at the position. This causes the EnvBuilder to crash.

`refinable_lvalue` just calls `expression` and also returns the refinement key-- since we've already called `expression` let's just pull out the refinement key instead!

New numbers:
`Passed: 182, Failed: 243, Skipped: 273, Errored: 44`

Changelog: [Internal]

Reviewed By: mvitousek

Differential Revision: D31588835

fbshipit-source-id: 89e483679e9b8dd88be98c37c2885e8a9ea9b9b4