default export declarations should use declared type
Summary:
Like in D30814150 (https://github.com/facebook/flow/commit/2857b8323690ee52bdafcc00d500cc92b438799b), the exported value should use the declared type and not the flow-sensitive type. In these positions they should be equivalent in the old env, but in the new_env this declaration is not a read of the variable so we must look at its providers.

Test results with new env:
`Passed: 182, Failed: 228, Skipped: 273, Errored: 59`

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D30820726

fbshipit-source-id: 5e9985c2ae3b393b2a094c2bb06be68e801f079f