get_var_declared_type should query providers
Summary:
Title. Loc_env.find_write is going to query the EnvBuilder for a read, aka the flow-specific type of the variable at that location, but get_var_declared_type is called on the declaration of the variable, so the providers should be queried instead.

New test results:

`Passed: 182, Failed: 222, Skipped: 273, Errored: 65`

Changelog: [internal}

Reviewed By: mvitousek

Differential Revision: D31696159

fbshipit-source-id: c64bd723a1bad238d55f4bdc85dd77b59d1d7687