unreachable reads in env builder
Summary:
The EnvBuilder throws an exception when the control-flow of a program ends in a definitive abrupt completion (continue, break, return, throw). This means that in code like:
```
let x = 3;
throw new Error();
x;
```
the EnvBuilder will never see the read of x on the third line. Statement.ml still traverses these reads in order to build the typed ast, which means the EnvBuilder must be able to answer queries about variables in unreachable code.

I came up with two solutions to handle this issue:
1. Treat reads that have no entry as unreachable
2. Explicitly mark unreachable variables as unreachable

The former is far simpler to implement than the latter, but this diff implements the latter. Why?

We want every variable read to have an entry in the EnvBuilder's output, and the absence of such an entry is an unacceptable state for us, since it'll type the variable as `empty` and silently hide the bug. We should try to avoid that situation at all costs. The second approach affords us the ability to discriminate between cases in which the EnvBuilder's output is incomplete vs. when the lookup location used in statement.ml is incorrect.

This diff adds another pass over the AST after the EnvBuilder is done running. We mark any variables that do not have an entry in the EnvBuilder's output as unreachable, and we convert that into an EmptyT when querying the environment.

New numbers:
`Passed: 183, Failed: 233, Skipped: 273, Errored: 53`

Changelog: [Internal]

Reviewed By: mvitousek

Differential Revision: D31587032

fbshipit-source-id: b3ca157a481985b8baf89a8ddc96214a0feb19a9