fix unnecessary recheck cancellation when multiple requests trigger prioritized rechecks
Summary:
### Bug

Suppose file A depends on B and B is unchecked. There's an unrelated recheck for C going on. The IDE fires 2 commands on A, say a hover and a definition. The hover request notices it needs to check B, pushes B onto the recheck stream and cancels the hover request (to be restarted after the recheck). The recheck of C is canceled, C is pushed back into the pending recheck workload, and a priority recheck of B is started. While the priority recheck of B is happening, we process the definition request, which also notices it needs to check B, pushes B onto the recheck stream and cancels the definition request.

Each time we push onto the recheck stream, we check to see if we should unblock `wait_for_updates_for_recheck` by calling `recheck_workload_is_empty`. However, because *C* is already in the recheck stream (we pushed it back there), `recheck_workload_is_empty` is false, so `wait_for_updates_for_recheck` unblocks incorrectly.

### Fix

`wait_for_updates_for_recheck` should only unblock if there are *updates* -- if the workload changed. priority rechecks changed an invariant: previously, anything waiting was an update; now that we requeue stuff, it might not be empty.

So, I replaced `recheck_workload_is_empty` with `workload_changed`, which actually checks to see if the new messages on the workload stream actually request anything new.

Reviewed By: samwgoldman

Differential Revision: D31783885

fbshipit-source-id: 88a397dc345e4250b95c9196d8feb80e58a4f96b