dismiss LSP interaction tracking in dismiss_tracks
Summary:
Our logging shows a large number of samples where `ux="Timeout"`, which is meant to indicate that a request has been outstanding without receiving a response for 10 minutes. I looked into where those are coming from and I found that up to 75% of them are from times where the server has intentionally restarted, e.g. when a lib file has been modified.

When the server dies to restart like this, the monitor closes the persistent connection to the LSP process and then the LSP process reconnects. When the LSP process gets kicked off the connection like this, `dismiss_tracks` sends cancellations to the client for any outstanding requests, but there was a bug where it didn't remove those requests from `LspInteraction`, so those requests would get logged as Timeouts 10 minutes later even though the LSP/server is healthy and those requests were responded to. This diff makes it so that `dismiss_tracks` dismisses tracks in `LspInteraction` as well, which should make it so that more (hopefully all) of the Timeouts that show up in our logging will be legit.

Reviewed By: mroch

Differential Revision: D31862776

fbshipit-source-id: 7169a68eb63fa1fc38d7b10eef7a42887e37709b