Simplify Scope.havoc_all_refis by dropping unused `?name` param
This parameter has never actually been passed since this function was
introduced.  Rather, the function has only ever been used in the mode
that its name suggests, to havoc *all* the refis in the scope.

By dropping that parameter, we can make the implementation very
simple, helping make it easier to reason about.  Because this makes
its interface simpler, it also lets us write a clean, concise comment
documenting what it's for.