Move ssa/scope/env analysis to separate folder
Summary:
It's been getting a little silly storing the env_builder and friends in parser_utils; and similarly, while the ssa_builder and friends are maybe relevant to parser utils, they're also used for a lot more than that. This diff pulls the scope/ssa builders into a new /src/analysis directory, intended for containing static analysis passes that are pre- or non-typechecking. This directory also contains a subfolder /env_builder that contains the analyses that power the new_env. Finally, env_builder.ml is renamed to name_resolver.ml.

Changelog: [internal]

Reviewed By: mroch, jbrown215

Differential Revision: D31842694

fbshipit-source-id: ab0ec02b9a1615cc47f55dc7a51b92c3c03d7c6d