fix referencing 'this' in a JSX title member
Summary:
Functions in statement.ml that handle JSX titles (like `jsx_title_member_to_expression`)
often convert `Ast.JSX.MemberExpression` to `Ast.Expressions` so that they can
be processed using the more general `expression` function. The case of
`Ast.JSX.MemberExpression (Identifier.name = "this")`, however, was not converted to
`Ast.Expression.This`, but `Ast.Expression.Identifer name = "this"`. This would cause
issues with typing down the line, because the latter is considered a type in the environments,
instead of an internal value.

This diff fixes this by converting `MemberExpression (Identifier name = "this")` to
`Ast.Expression.This` in `jsx_title_member_to_expression member`.

Changelog: [fix] referencing 'this' in a JSX title member

Reviewed By: samwgoldman

Differential Revision: D31929912

fbshipit-source-id: a1ac573e87acaa8754eb0d4f0735dec815501d38