[env-builder] Use havoc from env entry rather than querying providers for hoisted vals
Summary:
The `havoc` field on an environment entry directly records what the entry should be if it's havoced, e.g. by being hoisted. This diff changes the behavior of reading hoisted variables to read from the havoc field, rather than directly querying the providers. Not only is this simpler, it also handles refined globals in a way that I don't think the orignal solution did

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D31888225

fbshipit-source-id: ffe7ad375edb9afd6944beb66afac1a910686955