Fix export binding exception
Summary:
For statements like `export class C {}`, the signature builder uses a compact
representation, `ExportBinding offs` where `offs` is an offset into local defs
table.

However, in cases where the exported declaration is a duplicate binding error,
this led to unintended behavior, and sometimes crashes. The code that creates
the `ExportBinding` value expects to find the binding, but duplicate bindings
can not be added to the scope.

If the exported declaration duplicates another local definition, we will create
a `ExportBinding` value, but pointing to the previous binding. This is perhaps
unintended. If we find a remote binding instead of a local binding, then we will
crash on a `failwith`.

This diff adds a callback parameter to the binding functions which are used in
conjunction with exports. When the binding operation succeeds, it will invoke the
callback with the correct local def offset. If the binding operation fails, the
callback is not invoked.

We register the export from within this callback, causing duplicate binding
errors to be excluded from the module signature.

Changelog: [fix] Fix handling of duplicate bindings in signature builder

Reviewed By: mroch

Differential Revision: D31873554

fbshipit-source-id: bcf4389ffac1ac837e4600203f818d13ef2fdf7d