Initial work on type aliases name resolution
Summary:
This diff is the beginning of work on types within the env builder. For the most part, types don't need distinct handling within the env builder--they use the same `write` kinds as variables, etc. However, types are (at least in the new env--there are violations of this in the current env) constlike, in that havocing doesn't affect them, and we emit name-already-bound errors when they're redefined. This diff handles that by changing what val entries are generated in the env builder for type aliases--they're created as always pointing just to their declaration, even when havoeced.

We also now skip calls to `type_identifier_declaration` rather than treating them as reads (see D31887693). Only type identifiers on the right will be treated as reads. We don't need to do anything when they're on the left, because we already know what they always refer to.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D31888274

fbshipit-source-id: 18a092e806c53947aa0fc3e8d03d7ae6b9184bc8