Fix annotation information in fold_bindings
Summary:
Previously, `fold_bindings_of_pattern` (which is called by `fold_bindings_of_variable_declarations`) passed an annot into its callback based on the identifier pattern node. However, this is incorrect, since an annotation like this: ```let [ x: number ] = ...``` is ignored by flow; the correct pattern is to have a single annotation on the overall pattern. This resulted in cases like `let { x } : T` being considered to not have an annotation on `x`.

This fixes some behavior for constrained writes, but also improves the behavior of the local inference annotation requirement, as seen in the updated test case

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31809640

fbshipit-source-id: dbac0b227b2ea3f5c5b9cc6385cdeebea6827853