Propagate annotation information through parameters
Summary:
It's important for the purposes of constrained writes to know whether a function parameter is annotated or inferred. However, previously we treated all parameters as inferred. This diff threads information about the presence or absence of a parameter through the function parameter signature building code, and then uses it when binding the parameter.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31811728

fbshipit-source-id: d0027cb3486ed92612ad81e7b5f79c458facff40