Fix constrained writes for catch params and class expressions
Summary:
Previously when we tried to write to a catch param or a class expression name, we would get a constrained write error referring to the variable as having an unknown type, which meant that the type had no lower bounds when we resolved it. This fixes both cases; catch params are now treated like regular params and class names are treated as annotated.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31816574

fbshipit-source-id: 9ecc32ea65da5a90c7184d4933b94b5adaa72cdc