Delete derivable from reasons, dead with classic mode
This flag was used by classic mode for reporting missing-annotation
errors.  Since classic mode was deleted in cc78534ed / D25721686,
no remaining code has consulted this flag.

Specifically: the getter is_derivable_reason had one remaining caller,
which used it only to propagate the derivable bit to a new reason.
Other than that getter, nothing else in reason.ml looked at this flag
except to propagate it to a new reason, and no other code can see it.

Because the flag was unused, we can delete it.  This makes
derivable_reason the identity function, so delete that too; and that
makes Tvar.mk_decidable_where equivalent to Tvar.mk_where, so delete
it in favor of the latter.

(Looking at one of the other derivable_reason callers, builtin_reason,
I notice it's also dead.  But that one, I'm not quite sure if it's
permanently dead or might gain a caller later, so leave it for now.)