Add codemod heuristic to use type aliases in inferred unions
Summary:
This diff adds a heuristic to our codemod scripts that tries to
improve union types. If there is a single type alias in a branch of the union,
the updated type will only include that type alias. This avoids cases where a
bunch of literals are passed in and Flow's cannot tell they are part of the
type alias.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31913925

fbshipit-source-id: 13f7f13f34f2057db2dfb7bb29c808be30321bda