Make the Typeof ast node only allow qualified names, rather than arbitrary types
Summary:
It's unfortunate that currently the ast node within a `typeof` type expression is a `Ast.Type.t`. For one, it's a very different structure from what we actually allow--if `typeof`'s argument is anything but a generic (qualified or unqualified) with no type arguments we raise an error when converting to a `Type.t` in `type_annotation.ml`; for another, it means that when we traverse the AST we perceive the nodes within `typeof` as type-level identifiers rather than value identifiers.

This diff restructures the AST so that `typeof` targets can only be qualified names. The qualified names used are bespoke to `typeof`, rather than reusing `Generic.t`, in order to ban type arguments (allowed for `Generic.t`) and to allow identifiers with `typeof` to be visited as variable names, rather than type names.

Reviewed By: samwgoldman

Differential Revision: D31981311

fbshipit-source-id: 5cb82e0f15c8cae4882c7404f9ff31c7b1892cfe