fix flow-remove-types tests on node 16
Summary:
this test was passing `--code-comments` to node. this is a v8 flag that comments the assembly code when dumping that, which we're not doing in this test. the flag appears to have been removed in whatever version of v8 ships with node 16. it was just an arbitrary choice, so I changed it to a different flag.

also added a negative test, to make sure the flags actually get passed to node and not dropped.

Changelog: [internal]

Reviewed By: yungsters

Differential Revision: D32123451

fbshipit-source-id: 004139b7b1629a4408701df1a16af20af8bae6fe