remove `flow suggest` and `flow autofix suggest` commands
Summary:
the `flow suggest` command is a legacy API that attempts to insert inferred annotations on all function/method params and return types. it was also aliased as `flow autofix suggest`.

instead, this should be a codemod and/or a code action, using the newer `Insert_type` machinery. the latter handles inserting imports and has more advanced normalizer settings.

Changelog: [misc] removed the `flow suggest` command

Reviewed By: samwgoldman, panagosg7

Differential Revision: D32102700

fbshipit-source-id: 804543c89fb01205e6a62d1973dcecefa7ca36d6