Scopes for type parameters
Summary:
This diff builds support for recognizing the scopes of type parameters in the scope_builder, which enables them to be used by the env_builder.

The scoping for type parameters is fairly straightforward: in declaration like `<X: A = I, Y: B = J>`, `X` is not in scope for `A` or `I`, but is in scope for `B` and `J`. `Y` isn't in scope for anything (except presumably the function body or whatever that these parameters apply to). So, in the scope builder, for each type parameter we see, we explore its upper bound and default, and then we create a new scope that the parameter lives in, and then go to the next parameter. Once we've looked at all type parameters, we can explore the function body/class body/RHS of the type definition with all parameters in scope.

One tricky bit here is the scoping for a case like `class C<X> extends T<Y>`. In this case, `T` is not in the scope of any type parameters, but `Y` *is* in scope for all the type parameters. This means that we can write `type X<S> = S; class C<X> extends X<X>` and have the different reads of `X` in `X<X>` refer to different things! To handle this we need to split up what scope we explore the different parts of the `extends` clause in (and likewise `implements` and `mixins`).

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D31889246

fbshipit-source-id: 2ccc1d7a257ac2ed81df419b8323878bf18c3338