Initial version of name_def analysis for fully-resolved environment
Summary:
This diff gets us started on building the fully-resolved environment. The name_def module explores ASTs and builds a map from name locations--which should be identical to the locations used as keys in the env_builder--to the AST nodes will be analyzed to figure out the type of the name.

The result of this analysis will be used to create a fully resolved environment before doing any more work. We'll be able to figure out the right order to resolve the variables in the environment by exploring the AST nodes produced in this map, figuring out the dependencies within them using the env_builder, and topologically sorting.

The target of this mapping is a new data structure that contains the AST nodes directly, but also information about what kind of node it is, in order to decide which functions from statement.ml will ultimately be used to type it. In addition, to handle destructuring, we also use a `selector` type (very similar to the `selector` of type.ml) to decide how a single RHS of an assignment/declaration will define the types of multiple, destructured LHS variables. We generate these paths using a tool based directly on `destructuring.ml` but simplified greatly.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31905093

fbshipit-source-id: 6db3700e66eaef57b64f0627cdfc3d5b5345389d