More cases of type aliases
Summary:
This fixes a few issues with types in the env builder. First, we were originally skipping the read of `T` in `var x: T`; this diff fixes that by exploring the annotation on uninitialized variable declarations.

Second, when we see a qualified type like `React.ComponentType`, we'll now generate a read only of `React`. (Previously, this was generating a read of `React` and `ComponentType`, which is clearly wrong. At present we're not doing anything heap-refinement-like on types, and we probably shouldn't going forwards either.

This change also affects the dead_code mapper, so that it doesn't see `ComponentType` and think it's unreachable code.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31888408

fbshipit-source-id: 772b9c6c1766c3329018936b792daa5e4ed1de20