don't return code actions on non-@flow files
Summary:
code actions require types, which we should not try to compute on non-flow files. non-flow files are missing flow (unless `all=true`) or are ignored.

previously, we would try anyway, which could lead to incorrect results or bad perf. for example, if the file is ignored because it's actually part of a different project, and we typecheck it as part of the current project, then the results are unspecified; files with no annotations, or codegen, or compiled code, all are likely to be expensive to infer.

note that we already use the same `check_that_we_care_about_this_file` to avoid returning live errors or coverage for these files. we should do this check for all APIs.

Reviewed By: samwgoldman

Differential Revision: D32136352

fbshipit-source-id: 4d81bd27bc9c6ca44644225a27f023471a78a572