[annot-inf] handle elab_t catch-all case
Summary:
The catch-all case of `elab_t` in annotation inference should not be reachable,
since constraints are generated in a controlled way in type_sig_merge.

As with other parts of this file, we raise these errors just like Flow_js would, even
though these are not recorded in the active Context.t, to enable future uses of this
module in contexts other than new-merge.

In practice the only pattern that triggers this codepath is an export of the form
```
export type T = T.A | T.B;
```
which is already a Flow error.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D32123995

fbshipit-source-id: 67c905ed8a7f4c90ec4a506c817139a806e8b088