[annot-inf] OpenT indirection in Annot_GetPropT
Summary:
This diff adds an OpenT indirection to the result of get-prop in annot. inf. The
reason for this is purely to match the behavior of Flow_js.

One example where this discrepancy would be obvious is around `$Values` on frozen objects:
https://github.com/facebook/flow/blob/8c3825a1be188e9ade4ad4ed515361bb28c65d8a/src/typing/flow_js.ml#L1744-L1755

Due to this transformation, the following is an error
```
const FrozenSuite = Object.freeze({ DIAMONDS: 'Diamonds' });
('Hearts': $Values<typeof FrozenSuite>); // Error
```
but this is not
```
const IndirectFrozenSuite = Object.freeze({
  DIAMONDS: FrozenSuite.DIAMONDS,
});
('Hearts': $Values<typeof IndirectFrozenSuite>); // no errors
```
This indirection preserves this behavior to avoid user confusion.

Reviewed By: samwgoldman

Differential Revision: D32164226

fbshipit-source-id: ec0d0721af718ec3904c7e1cd7e07fab64ddc695