fix exception when live errors get requeued
Summary:
we keep track of the latest live error request and cancel all the older ones. we do this by keeping a mutable map from uri -> latest request metadata. we clear it when we've handled the latest request to avoid leaks.

when there's an unchecked dependency, we cancel the live error request, do a recheck, and then restart the live error request. we *also* trigger another live error request whenever a recheck ends.

if this second request is handled successfully before the first is requeued, then the 2nd request removes `uri` from the map (correctly, because it is the latest request). then the first request is requeued and when it runs, the `SMap.find` throws `Not_found`.

fix: use `find_opt`. if it's `None` that means that something newer already came and erased it, so cancel the request.

Changelog: [internal]

Reviewed By: vrama628

Differential Revision: D32197786

fbshipit-source-id: 2d3f184def34e586e8366919fe505884834709fb