only provide hovers / type-at-pos on @flow files
Summary:
type-at-pos requires types, which we should not try to compute on non-flow files. non-flow files are missing flow (unless all=true) or are ignored.

previously, we would try anyway, which could lead to incorrect results or bad perf. for example, if the file is ignored because it's actually part of a different project, and we typecheck it as part of the current project, then the results are unspecified; files with no annotations, or codegen, or compiled code, all are likely to be expensive to infer.

this diff calls `check_that_we_care_about_this_file` in `infer_type`, which backs hover LSP requests and `flow type-at-pos`.

Changelog: [fix] No longer attempt to provide hover or `flow type-at-pos` on non-Flow files

Reviewed By: samwgoldman

Differential Revision: D32239134

fbshipit-source-id: bf40c501618a1a7b6b1a8192f2e2761b92902ac2