[indexed-access] Delete indexed_access option, it's always enabled now
Summary:
The `indexed_access` option has defaulted to `true` for a while now. Now that all of our repos support indexed access types, we can fully remove the flag and have it always be enabled.

Changelog: [feature] Remove `indexed_access` option, it's always enabled now (was on by default before if the option was omitted)

Reviewed By: mroch

Differential Revision: D32240135

fbshipit-source-id: 70f8f7880c5d2373397d2496c19c411b98ce5712