only provide get-def on @flow files
Summary:
get-def (`flow get-def` and `textDocument/definition`) requires types, which we should not try to compute on non-flow files. non-flow files are missing flow (unless all=true) or are ignored.

previously, we would try anyway, which could lead to incorrect results or bad perf. for example, if the file is ignored because it's actually part of a different project, and we typecheck it as part of the current project, then the results are unspecified; files with no annotations, or codegen, or compiled code, all are likely to be expensive to infer.

this diff calls `check_that_we_care_about_this_file` in `get_def`, which backs definition LSP requests and `flow get-def`.

Changelog: [fix] No longer attempt to provide definition or `flow get-def` on non-Flow files

Reviewed By: samwgoldman

Differential Revision: D32240404

fbshipit-source-id: 2ada413627a439f952a300e0834e15e7e4595ac6