[tool] add LSP snapshot testing
Summary:
adds `verifyLSPMessageSnapshot`, which takes the path to a file whose contents is used as the expected contents of all of the LSP messages in the step (except any ignored ones). this is very similar to `verifyAllLSPMessagesInStep`, except the latter takes an array of message objects directly instead of reading from a file.

`tool record` is also updated to update the snapshot file.

Changelog: [internal]

Reviewed By: vrama628

Differential Revision: D32126190

fbshipit-source-id: 730662ca2cedb6371ae2dc71cd565a42924bd311