remove Lwt from most LSP commands
Summary:
using Lwt allows an LSP command to run in parallel with a recheck, but it also allows commands to be interrupted by slow recheck work happening in the main process.

since we run one command at a time, at least they're not interrupting each other -- only a recheck will starve a command. since commands take priority over rechecks, we never want a command to be interrupted.

ideally, nothing would ever do slow work in the main process, and we'll explore moving rechecks or commands or both into separate processes.

but, interestingly, the ONLY thing we use Lwt in the command handlers for is memory profiling! the memory profiler runs a background thread that profiles memory every second. But I should say "attempts to run," because the functions being run don't themselves use Lwt or do I/O, so they never yield, so the timer thread never runs. Further, we're not actively using this memory profiling.

so, this diff strips the memory profiling, replacing it with just `Profiling_js.with_timer`, and then removes all the now-unnecessary Lwt.

Changelog: [misc] Changed scheduling of LSP commands in parallel with typechecking, to prioritize commands.

Reviewed By: samwgoldman

Differential Revision: D32241501

fbshipit-source-id: 005700181ed5f5b9abda367eea55c78ad792d747