[indexed-access] Always output indexed access in ty normalizer for ElementType and PropertyType
Summary:
Now that Indexed Access is on by default, always output it from the ty normalizer for `ElementType` and `PropertyType`. This will encourage people to discover and use the feature, and prepares for the eventual deletion of `ElementType` and `PropertyType` some point later.

Changelog: [feature] When displaying types, always use indexed access types for `$ElementType` and `$PropertyType` utilities

Reviewed By: panagosg7

Differential Revision: D32240134

fbshipit-source-id: 4772a117fc33ce2f96f9baf91582fa9f6a5cd601