improve LSP logging in tool
Summary: in the `test.log` output of our `tool` tests, we manually log all the activity that happens on the LSP connection, but we don't currently include the LSP request IDs in most cases. `tool` is proving to be useful for reproducing and debugging specific scenarios, and having the IDs of the LSP requests in the logs will make it even better for that purpose. This change basically makes it so that `tool`'s LSP log will look like VSCode's.

Reviewed By: mroch

Differential Revision: D31872244

fbshipit-source-id: c026952e87a165aadff326a82340695c2e725bb7