[cleanup] avoid a CheckedSet.all during recheck
Summary:
more efficient to just directly check membership in the two collections than it is to convert `CheckedSet` to a `FilenameSet`, then union it.

but more importantly, all these different sets of files we keep track of during recheck are impossible to understand, so this gets rid of one and is no harder to understand because it now literally matches the comment.

Reviewed By: samwgoldman

Differential Revision: D32416167

fbshipit-source-id: 093cbed2e2bebf646d55aa09519eb0258e9d23fc